wallet_service/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
Django==5.1.3
django-environ==0.11.2
django-redis==5.4.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.2
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
locust==2.32.3
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pytest==7.4.0
pytest-django==4.9.0
pytest-lazy-fixture==0.6.3
python-dateutil==2.9.0.post0
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.1.1
